requirements.txt
finnhub-python
requests
python-dotenv
fastapi
uvicorn
pydantic
python-dotenv
openai
pyyaml
reportlab
numpy
pandas
reportlab
pymupdf
pyPDF2
jinja2
azure-storage-blob
sec-api
python-docx
azure-search-documents
azure-core
tiktoken
langchain-community
langchain-openai
pypdf 
faiss-cpu
